==> Photodegradation/01 - Absorbance_spectro.R <==
# LapierreLab
# Author: Philippe Maisonneuve
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
# This script corrects absorbance data for water absorbance and cuvette length
# and genereates a unified .csv file of all the scans contained into a data file
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #

#install.packages("BiocManager")
#install("limma")
library("BiocManager")
library(limma)

# Raw absorbance data generated by the Shimadzu are stored as .txt 
# files in a "CDOM" folder located in a folder "data" which will be
# set as the working drectory
setwd("~/Documents/Labo Lapierre/Maîtrise/R/Projects/PARAFAC 2019/data/CDOM")

# Create a list containing the names of all the files in the CDOM folder
data.files = list.files(path = "./")

# Create a data frame to store the corrected spectral data
output <- data.frame(matrix(0, nrow = 711, ncol = length(data.files)+1))
colnames(output)[1] <- "WL"
output[,1] <- c(190:900)
                  

for(i in 1:length(data.files))
{
  
  temp <- read.table(paste0(data.files[i]), header = T, skip = 1, sep = ",")
  temp.name <- (paste0(data.files[i]))
  
  # Substracting the absorbance of water
  a690 <- temp[temp$Wavelength.nm.==690,2]
  corr690 <- data.frame(temp$Abs. - a690)
  
  #Removing negative values (no physical sense)
  for(p in 1:nrow(corr690))
  {
    if (corr690[p,1] < 0)
      
      corr690[p,1] <- 0    
  }
  
  
  
  # Dividing by the pathlength of the cuvette
  Ag <- data.frame(corr690 / 0.01)
  

  #Converting from absorbance to absorption coefficient (Vahatalo et al. 2000)
  Ag <- data.frame(Ag*log(10))
  
  #Storing the transformed spectra
  data.files[i] <- removeExt(data.files[i], sep = ".")
  colnames(output)[i+1] <- data.files[i]
  rownames(output) <- temp$Wavelength.nm.
  output[,i+1] <- Ag
  
}

y <- output[output$WL==440,]
#Selecting only a specific wavelength (e.g. a254, a440)
#output <- output[output$WL==254,]
output <- output[output$WL==440,]

setwd("~/Documents/Labo Lapierre/Maîtrise/Data/2019/MSL2019 - CDOM")
write.csv(x = output, file = "Absorbance_MSL2019.csv", fileEncoding = "UTF-8")





















for(i in 1:length(data.files))
{
  
  temp <- read.table(paste0(data.files[i]), skip = 1, header = T, sep = ",")


      output[i,1] <- data.files[i]
      output[i,2] <- (temp[temp$Wavelength.nm.==254,2]) * 2.303 / 0.01
      output[i,3] <- (temp[temp$Wavelength.nm.==440,2]) * 2.303 / 0.01
      
}
  
  write.csv(x = output, file = "Absorbance_MSL2019.csv", fileEncoding = "UTF-8")
  
  #SpectralSlope <- function(data.file = "data", wl0 = 190, From = 250, To = 600, By = 1, skip = 1)
  {
    
    #create sequence of desired wavelengths
    wl.x = seq(From, To, By)
    #Create the output matrix
    output = matrix(0,ncol = 4, nrow = 1)
    colnames(output) = c("Intercept", "Slope", "R2", "K")
    rownames(output) = data.files[i]
    
    #Compute the exponential fit for each CDOM file
    
    Abs = read.table((data.file), skip = skip, header = skip + 1, sep=",")
    WL = Abs[,1]
    a0 = Abs[which(Abs[,1] == wl0),2]
    Abs.y = sapply(wl.x, function(x){return(subset(Abs[,2], WL == x))})
    nls.temp <- nls(Abs.y ~ a0 * exp(-S * (wl.x - wl0)) + K, 
                    start = list(a0 = a0, S = 0.02, K = 0.02), #a0 is bounded between 0 and the highest observed value
                    lower = list(a0 = 0, S = 0, K = 0), #S is bounded between 0 and 1 otherwise the function grows
                    upper = list(a0 = max(Abs.y), S = 1, K = 3),
                    control = list(maxiter = 200, warnOnly = T),
                    algorithm = "port") #K is theoritically unbound [-Inf:Inf], but such values make no sense in this context. We've set the boundary to 0 and 3, which is the limit for absorption correction used with EEMs ()
    
    R2 <- 1 - sum((Abs.y - predict(nls.temp))^2) / (length(Abs.y) * var(Abs.y)) #Denominator is sum(y-mean(y))² which is variance(y) times length(y)
    
    output[1,1] = coef(nls.temp)[1]
    output[1,2] = coef(nls.temp)[2]
    output[1,3] = R2
    output[1,4] = coef(nls.temp)[3]
    
    class(output) = "SpectralSlope"
    
    #Return the result
    return(output)
  }
  
 
  
  
  
  ############### RSpectroAbs ###########
  ###### GitHub de Simon Belanger (belasi01) #####
  
  if (!require("devtools")) install.packages("devtools")
  library("devtools") 
  devtools::install_github("belasi01/Riops")
  devtools::install_github("belasi01/RSpectroAbs")
  library("RspectroAbs")
  library(RspectroAbs)
  
  #list CDOM files as csv
  log_file <- list.files(path="./CDOM")
  write.csv(files.list, file = "log_file.csv") 
  #need to add: ID    	Station    	 Depth   	pathlength   	Ag.good 	DilutionFactor
  #ID is over the .txt files.
  #Station is same as ID
  #Depht is NA
  #pathlength is 0.01
  #Ag.good and Dilution is 1
  #save as text also
  
  setwd("~/UdeM Doctorat/Donn?es/parafac/TotalCDOMFiles")
  #liste fichier dans TotalCDOMFiles
  CDOMList<-list.files(path="C:/Users/Steph/Documents/UdeM Doctorat/Donn?es/parafac/TotalCDOMFiles")
  #have folder "csv" en minuscule in parafac file
  
  #rewrite .txt as .csv:
  
  for(i in 1:length(CDOMList)){
    
    #reorder .txt files to have highest nm (900) on top
    temp <- read.table(paste0("./", CDOMList[i]), skip = 1, 
                       header = T, sep = ",")
    sample <- strsplit(CDOMList[i],".txt")[[1]]
    write.csv(x = temp, file = paste0("../csv/",sample,".Sample.Raw.csv"),
              fileEncoding = "UTF-8")
  }
  
  
  RspectroAbs::run.process.Ag.batch(log.file = "../C.txt", 
                                    data.path = "C:/Users/Steph/Documents/UdeM Doctorat/Donn?es/parafac")
  
  
  #Error in fit_exponential() when trying to fit. Check your data.
  #singular gradient matrix at initial parameter estimatesError in
  #fit_exponential() when trying to fit. Check your data.
  
  ### checking data:
  
  setwd("~/UdeM Doctorat/Donn?es/parafac/TotalCDOMFiles")
  
  abs<-read.csv("P1-89.txt", skip=2, header=F,dec=".") 
  plot(V2~V1,abs)
  rich<-read.csv("C:/Users/Steph/Desktop/400133.Sample.Raw.csv")
  plot(A~nm,rich)
