==> Photodegradation/03 - Kd_Attenuation_PUV.R <==
# LapierreLab
# Author: Philippe Maisonneuve
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
# This script calculates light attenuation rates (Kd)
# from data files generated by the PAR-UV probe (Bertolo Lab, UQTR)
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #

#install.packages("dplyr")
library(dplyr)
#install.packages("zoo")
library(zoo)

#NaRV.omit function
#install.packages("IDPmisc")
library(IDPmisc)


# Calibrated data generated by the PAR-UV probe are stored as .csv 
# files in a "CDOM" folder located in a folder "data" which will be
# set as the working drectory
setwd("~/Documents/Labo Lapierre/Maîtrise/Data/2019/PUV/")

# Create a list containing the names of all the files in the CDOM folder
data.files = list.files(path = "./csv/")

# Create a matrix to store Kd values, R2, and maximum depth at each site
output = matrix(0,ncol = 3, nrow = length(data.files))
colnames(output) = c("Kd443", "R2.Kd443", "Max_Depth") #"Kd313", "R2.Kd313" #Kd320", "R2.Kd320","Kd340", "R2.Kd340","Kd443", "R2.Kd443","Kd550", "R2.Kd550",  "Max_depth")
sites.names <- sub('.*\\_', '',data.files) 
sites.names <- sub("\\..*","",sites.names)
rownames(output) <- sites.names


pdf(file="Kd313.pdf")
for(i in 1:length(data.files))
{
header <- read.table(paste0('./csv/',data.files[i]), nrows = 1, header = FALSE, sep =',', stringsAsFactors = FALSE)
data <- read.table(paste0('./csv/',data.files[i]), skip = 2, header = FALSE, sep =',')
colnames(data) <- unlist(header)
data$Depth <- as.numeric(as.character(data$Depth))
a <- cbind(data$Depth, log(data$EdZ340))
a <- NaRV.omit(a)
colnames(a) <- c("Z", "Iz")

f <- function (d) {
  m <- lm(Iz~Z, as.data.frame(d))
  return(summary(m)$r.squared)
}


# Apply function to data with a moving window of 10 observations
# and correlation treshold of R^2>0.95
r2 <- rollapply(a, 15, f, by.column=F)
id.points <- which(r2 > 0.95, r2)
RES <- data.frame(a[id.points,])

# Plot conserved (in red) and discarded points 
# with linear trend line and graphical parameters
plot(a, main = sites.names[i])
points(RES,pch=15,col="red")
lm.Kd <- lm(Iz~Z,RES)
abline(lm(Iz~Z,RES),col="blue")
legend("topright", legend = paste("R^2 =",
                                  round(summary(lm.Kd)$adj.r.squared, digits = 4),
                                  "\nKd = ",
                                  round(coef(lm.Kd)[2], digits = 4)))

output[i,1] <- round(coef(lm.Kd)[2], digits = 4)
output[i,2] <- round(summary(lm.Kd)$adj.r.squared, digits = 4)

# Z0 is the depth at which we start to see light attenuation (the depth at which the instrument enters the water)  
# We need to substact this depth to the maximum depth to obtain the real depth
output[i,3] <- max(data$Depth, na.rm = TRUE) - RES[1,1]

}

dev.off() 

write.csv(x = output, file = "Kd313.csv", fileEncoding = "UTF-8")
