==> Photodegradation/04 - Average light dose_WaterColumn.R <==
# LapierreLab
# Author: Philippe Maisonneuve
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
# This script calculates the average areal light dose (in W*m^-2)
# in a given water column of depth z (m)
# Input: Irradiance profiles (μW/(cm^2*nm))
# Output: Average areal light dose (W*m^-2)  
## Variables  
## Ed0: Surface irradiance
## Kd_XXX: Attenuation coefficient of light at XXX nm

# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #


library(readxl)
#install.packages("dplyr")
library(dplyr)
#install.packages("zoo")
library(zoo)

#NaRV.omit function
#install.packages("IDPmisc")
library(IDPmisc)
library(readr)


# Calibrated data generated by the PAR-UV probe are stored as .csv 
# files in a "CDOM" folder located in a folder "data" which will be
# set as the working drectory
setwd("~/Documents/Labo Lapierre/Maîtrise/Data/2019/PUV/")

# Creating a list containing the names of all the files in the CDOM folder
data.files = list.files(path = "./csv/")

# Selecting the file of interest by adjusting the value of i
# (or go further in the script for the looped version)
# i = 1

output = matrix(0,ncol = 9, nrow = 40)
colnames(output) = c("Station", "avg_LD","avg_Daily_Energy", "Max_Depth", "ratio")

for(i in 1:40)
{
#Exporting and organizing the average surface irradiance data chosen to be TRA3 in script 06- Average daily dose_St-Lawrence
header <- read.table(paste0('./csv/',data.files[1]), nrows = 1, header = FALSE, sep =',', stringsAsFactors = FALSE)
data <- read.table(paste0('./csv/',data.files[1]), skip = 2, header = FALSE, sep =',')
sites.names <- sub('.*\\_', '',data.files) 
sites.names <- sub("\\..*","",sites.names)
sites.names <- sort(sites.names)
colnames(data) <- unlist(header)
data$Depth <- as.numeric(as.character(data$Depth))

# 1- Interpolating surface irradiance values (Ed0-λ) to generate a spectra from 313-450nm

  #Average Ed0-λ values for a given station i and store them
  Ed0_raw <- data.frame(matrix(0,ncol = 2,nrow = 5))
  colnames(Ed0_raw) = c("WL", "Ed0")
  Ed0_raw[,1]  <- c(313,320,340,443,550)
  Ed0_raw[1,2] <- mean(data$Ed0313, na.rm = TRUE)  
  Ed0_raw[2,2] <- mean(data$Ed0320, na.rm = TRUE)
  Ed0_raw[3,2] <- mean(data$Ed0340, na.rm = TRUE)
  Ed0_raw[4,2] <- mean(data$Ed0443, na.rm = TRUE)
  Ed0_raw[5,2] <- mean(data$Ed0550, na.rm = TRUE)
  plot((Ed0_raw$Ed0)~Ed0_raw$WL, xlab= "wavelength (nm)", ylab = "Ed0 (μW/(cm^2*nm))")
  #Convert to W/m^2
  Ed0_raw$Ed0 <- Ed0_raw$Ed0*(10^-6)*10000
  
  #Linear interpolation to 1-nm increments between 313-550nm
  approxSpectra <- data.frame(approx(Ed0_raw, xout = seq(313, 550, by = 1), method = "linear"))
  plot(approxSpectra, xlab= "wavelength (nm)", ylab = "Ed0 (W/(m^2*nm))")
  #Creating a function performing the linear interpolation
  Ed0_Spectrum <- approxfun(Ed0_raw, method = "linear") 
  integrate(Ed0_Spectrum, 313, 450) 
  
  #Storing the Ed0 values
  #output[i,2] <- (a$value/(10^6))*10000
  #output[i,1] <- sites.names[i]

#write.csv(x = output, file = "Ed0.csv", fileEncoding = "UTF-8")
#output.df <- data.frame(output)
#output.df$Ed0 <- as.numeric(output.df$Ed0)
#plot(output.df$Ed0)
#hist(output.df$Ed0)

  # Create a matrix to store average areal light doses and maximum depth at each site 
  #output = matrix(0,ncol = 2, nrow = length(data.files))
  #colnames(output) = c("AALD", "Max_Depth")
  #sites.names <- sub('.*\\_', '',data.files) 
  #sites.names <- sub("\\..*","",sites.names)
  #rownames(output) <- sites.names

# 2- Importing the Kd values
    
  Kd <- data.frame(read_excel("~/Documents/Labo Lapierre/Maîtrise/Data/2019/MSL2019_Data.xlsx"))  
  Kd <- cbind(Kd$Kd_313, Kd$Kd_320, Kd$Kd_340, Kd$Kd_443, Kd$Kd_550)
  WL <- c(313, 320, 340, 443, 550)
  Kd <- Kd[i,]
  Kd <- data.frame(rbind(WL,Kd))
  Kd <- t(Kd)
  
  #Linear interpolation to 1-nm increments between 313-550nm
  Kd_Spectrum <- data.frame(approx(Kd, xout = seq(313, 443, by = 1), method = "linear")) 
  plot(Kd_Spectrum, main = paste0("Kd - ",sites.names[i]))
  
  #Creating a function performing the linear interpolation of Kd (between 313-550nm)
  Kd_Spectrum <- approxfun(Kd, method = "linear")
  
# 3- Importing the maximum depth
  
  MSL2019_Data <- data.frame(read_excel("~/Documents/Labo Lapierre/Maîtrise/Data/2019/MSL2019_Data.xlsx"))
  Depths <- subset(MSL2019_Data, select = MaxDepth_PUV_corrected)
  z <- Depths[i,]


# 4 - Importing the CDOM spectra
  
  CDOM <- data.frame(read.csv("~/Documents/Labo Lapierre/Maîtrise/Data/2019/MSL2019 - CDOM/Absorbance_MSL2019.csv", sep= ","))
  CDOM_i <- CDOM[,which(colnames(CDOM)==paste0(sites.names[i],".T0"))]
  Spectrum_i <- data.frame(cbind(CDOM$WL, CDOM_i))
  #CDOM_T <- CDOM[,which(colnames(CDOM)==paste0(sites.names[i],".T0"))]
  #CDOM_T <- data.frame(cbind(CDOM$WL, CDOM_T))
  #CDOM_i <- (CDOM_P+CDOM_T)/2
  #Spectrum_i <- data.frame(cbind(CDOM$WL, CDOM_i))

  #Linear interpolation to 1-nm increments between 313-550nm
  #CDOM_T <- data.frame(approx(CDOM_T, xout = seq(190, 900, by = 1), method = "linear"))
  #CDOM_P <- data.frame(approx(CDOM_P, xout = seq(190, 900, by = 1), method = "linear"))
  #plot(CDOM_P, main="CDOM P", ylim=c(0,100))
  #plot(CDOM_T, main="CDOM T", ylim=c(0,100))
  
  #Creating a function performing the linear interpolation of Kd (between 313-550nm)
  CDOM_Spectrum <- approxfun(Spectrum_i, method = "linear")
  plot(CDOM_Spectrum, xlim=c(300,443))
  #lo <- loess(Spectrum_i$CDOM_i~Spectrum_i$V1)
  #plot(Spectrum_i$V1,Spectrum_i$CDOM_i)
  #lines(loess.smooth(Spectrum_i$V1,Spectrum_i$CDOM_i),col=8)
  
# 5 - Calculating the ratio between CDOM and Kd (indicator of what fraction of light is absorbed by CDOM)  
  
ratio <-  function(x)
{
  CDOM_Spectrum(x)/Kd_Spectrum(x)
}

plot(ratio, xlim=c(313,550), ylim=c(0,2), main = paste0(sites.names[i]))


# 6 - Resolving the integral (formula from Cory et al. 2013)
  
  f <- function(x)
  {
    Ed0_Spectrum(x)*(1-exp(-Kd_Spectrum(x)*z))*ratio(x)
  }
  
  
  
  plot(Ed0_Spectrum, xlim = c(313,550), ylim=c(0,1), main=sites.names[i])
  plot(CDOM_Spectrum, xlim = c(313,550), ylim=c(0,10), main=sites.names[i])
  plot(Kd_Spectrum, xlim = c(313,550), ylim=c(0,10), main=sites.names[i])
  
  plot(f, xlim = c(313,550), ylim=c(0,2), main=sites.names[i])
  # Average areal light dose (W/m^2)
  LD <- integrate(f, 313, 450, subdivisions = 2000)
  
  #Storing the final result
  output[i,1] <- sites.names[i]
  output[i,2] <- LD$value 
  output[i,3] <- LD$value*(86400) #Multiplied by the average time of day light per day during mission (see script 06)
  output[i,4] <- z
  output[i,5] <- ratio(313)
  output[i,6] <- ratio(320)
  output[i,7] <- ratio(340)
  output[i,8] <- ratio(443)
  output[i,9] <- ratio(550)

}
  

output <- data.frame(output)
output$ratio <- as.numeric(as.character(output$ratio))
output$X6 <- as.numeric(as.character(output$X6))
setwd("~/Documents/Labo Lapierre/Maîtrise/Data/2019")
write.csv(x = output, file = "Average_areal_light_dose_WaterColumn_T0.csv", fileEncoding = "UTF-8")

# 6- Calculating the AQY  
 
  install.packages("pracma")
  library(pracma)
  WL = 313 
  f_AQY <- function(c, d)
  {
    c*exp(-d*WL)
  }
  
  
  fminsearch(f_AQY, 0)
  
#for(i in 1:length(data.files)) 
  
  legend("bottomright", legend = paste("R^2 =",
                                                                         round(summary(lm1)$adj.r.squared, digits = 4)))
